refatorando codigo de criacao de contatos
