--- desafio3/src/button/adding_contact/create_contact.py
@@ -1,8 +1,0 @@
-def create_contact() -> tuple[str, str, str]:
-    contact_informations = list()
-    contact_informations.append(insert_name())
-    contact_informations.append(insert_phone())
-    contact_informations.append(insert_email())
-
-    print("Contato adicionado!")
-    return tuple(contact_informations)
