--- desafio3/src/button/adding_contact/insert_name.py
@@ -1,6 +1,0 @@
-def insert_name() -> str:
-    name = input("Nome: ")
-
-    if not name.replace(" ", "").isalpha():
-        raise ValueError("Insira apenas letras.")
-    return name
